v1.20.00
-[NEW] Cocos2d-JS support
-[FIX] Bug fixes

+++ External Cocos Helper Android Frameworks/src/sonar/systems/framework/SonarFrameworkFunctions.java
@@ -678,7 +678,7 @@ public class SonarFrameworkFunctions
 		}
 	}
 	
-	public static void SendGAEvent(final String category, final String action, final String label, long value)
+	public static void SendGAEvent(final String category, final String action, final String label, int value)
 	{
 		if(SonarFrameworkSettings.USE_GOOGLE_ANALYTICS)
 		{
